+++ CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0]
+### Added
+- floor level, floor certainty, altitude and bearing on location change
+- quality when calibration changes
+
+## [0.1.0] 2019-01-06
 ### Added
 - exposed location service status codes and calibration codes constants
 - regionEnter event, with a lot of floor map data